Add devstack-tools
devstack-tools is a port of some of the more complicated bash code in
devstack to python clis to simplify the maintenance of these. The
initial focus is modification of conf files, including the devstack
local.conf files. It is needed to support local.conf in the gate, and
as such will need to be ported back to all stable trees in
requirements.

Change-Id: I6ef05adebc9ecfb9710d07b71d9df42e4eef97f1
(cherry picked from commit 97ec7440258f4a43d05d03dd45f29acf2812f74a)