add blank lines to the constraints list to avoid merge conflicts
When we release several libraries that appear in the same portion of the
constraints list, we regularly have merge conflicts with those
individual constraint updates. This patch causes the automatic generator
to introduct blank lines between the constraints so that the individual
updates from the release jobs are less likely to appear to be modifying
the same section of the file, so there are fewer merge conflicts.

Change-Id: I36d0bae63711f3edeb448c21209c677895ca9f31
Signed-off-by: Doug Hellmann <doug@doughellmann.com>