sort by non-repeating prefixes of project names
We have a bunch of deliverables with the same prefix in their name
"oslo*" or "python-*". When we release several of those around the same
time, as often happens with the Oslo libraries, we end up with merge
conflicts for the separate patches that update the constraints
list. This patch changes the sort order of the constraints list to
ignore those common prefixes to try to spread similar names out around
the list to avoid conflicts.

Change-Id: Ica334f1afc72dfec1fd69b63806312691b8f1468
Signed-off-by: Doug Hellmann <doug@doughellmann.com>