Update patch set 1
Patch Set 1:

Matthew, I don't think there's any possibility of getting a synchronous mode in the new kafka client itself.  There isn't any client I've found that offers that.  What we would have to do is rewrite enough of Monasca to layer the reliability we want on top of the async mode that the kafka client provdes.

Gordon, I agree that the increased memory requirement is not a leak but it has nothing to do with sending data to Kafka.  The increased memory I observed is on the read side.  Consuming and processing identical data between the 0.9.5 client and the 1.3.1 client required 2x more memory in 1.3.1.

Patch-set: 1