Update patch set 1
Patch Set 1:

If using a different library is a viable option, which hasn't seemed to be the case in global requirements yet, I have a suggestion.  Monasca can fork the older version of kafka-python and embed it for now.  That'll let oslo.messaging move forward with whatever they need to do.  We'll update Monasca over time once we figure out how to make the async client reliable enough.  If we're going to be doing that sort of rewrite though we'd prefer to rewrite against a library that has higher throughput.  What we'd like to do is add a second Kafka client library to global requirements, one based off of librdkafka, that will let us chase the performance and stability targets we want and oslo.messaging can chase after the features they want in kafka-python.

Patch-set: 1