Update patch set 1
Patch Set 1:

oslo.messaging is under a stricter time constraint then you are I think.  Would it be possible for you to simply deny the kafka-python update when it shows up in your queue (edit the review to remove the kafka-python change).  This would allow us to update the kafka-python dep now for oslo.messaging to use.

Patch-set: 1