Update patch set 4
Patch Set 4:

Message looks better.

For code review, it looks like they are not passing their own gating.
They do have 84% coverage, which I think is good enough.
Their requirements (and test-requirements) are a superset of ours so should be fine.

This is much like the yaml option we had before (we have two yaml libs in requirements, speed was an issue as well).
I think it'll be fine, but the build error is worrying.

https://travis-ci.org/Parsely/pykafka

Patch-set: 4