Update patch set 6
Patch Set 6:

Yes, in a way.

SDK doesn't follow the general release schedules, but our 1.0 is slated to be done probably slightly after Ocata final, likely early March. In order to hit that, there are several APIs we'd like to run through a period of deprecation starting soon so that the affected users have time to catch up over the remaining couple of 0.9.x releases over the next few weeks.

Patch-set: 6