Update patch set 2
Patch Set 2:

hmm, I don't think, are allowed to add something once we've released, but this may be a special situation.  That this is to get closer to using requirements is a good thing though.  Also good is that it's tested as a dependant patch successfully.  I'll try and bring this up with the other cores.

Patch-set: 2