Update patch set 1
Patch Set 1:

I did the creation of the tag by hand as described in the project creators guide. The update constraint job of the corresponding release resulted in:

2017-01-31 10:33:15.590950 | + echo 'WARNING: Missing meta lines in 0.2.0 tag message,'
2017-01-31 10:33:15.590977 | WARNING: Missing meta lines in 0.2.0 tag message,
2017-01-31 10:33:15.590998 | + echo '         skipping announcement.'
2017-01-31 10:33:15.591018 |          skipping announcement.
2017-01-31 10:33:15.591031 | + echo
2017-01-31 10:33:15.591043 | 
2017-01-31 10:33:15.591069 | + echo 'Was the tag for 0.2.0 created with release.sh?'
2017-01-31 10:33:15.591092 | Was the tag for 0.2.0 created with release.sh?
2017-01-31 10:33:15.591106 | + exit 0

That's why it got not proposed automatically I guess. Is there anything that I need to do in addition to get the metadata added? The project is not managed by openstack/release so far... and not sure if it makes sense, as it is not under the openstack governance...

Patch-set: 1