Update patch set 1
Patch Set 1:

While no formal FFE has been applied for Jeffrey asked me to look at this on IRC[1]

I'm okay with taking this patch as it only affects requirements and murano projects.

So if we do an FFE on the mailing list it shoudl be quick to grant.

[1] http://eavesdrop.openstack.org/irclogs/%23openstack-release/%23openstack-release.2017-02-07.log.html#t2017-02-07T00:11:57

Patch-set: 1