Update patch set 3
Patch Set 3: Code-Review+2

The reasons for doing this seem sound. We will wait to branch the requirements repository until this patch has merged (though we don't need to wait for the related patches to merge in horizon and the other projects that use these packages because we'll get an update for them on their stable/ocata branch anyway).

Patch-set: 3
Label: Code-Review=+2