Update patch set 1
Patch Set 1: Code-Review-1

Looks like you also need to add this to upper-constraints.

I think this is a good thing to add though. I've seen this issue come up in reviews. If the checking for that can be automated, I'm sure it will take a lot of cleanup, but once there it would be good to have it enforced without depending on reviewers.

Patch-set: 1
Label: Code-Review=-1