Update patch set 2
Patch Set 2:

Tony, does [1] actually enforce anything, or is that just the template that should be followed? I know at least in Cinder, that definitely is not enforcing those guidelines, so often it ends up whether a reviewer notices/comments on it or not. When I saw this, I was hoping adding this would eliminate that need.

[1] http://docs.openstack.org/developer/hacking/#import-order-template

Patch-set: 2