Update patch set 2
Patch Set 2: Code-Review+2

> Tony, does [1] actually enforce anything, or is that just the
 > template that should be followed? I know at least in Cinder, that
 > definitely is not enforcing those guidelines, so often it ends up
 > whether a reviewer notices/comments on it or not. When I saw this,
 > I was hoping adding this would eliminate that need.
 > 
 > [1] http://docs.openstack.org/developer/hacking/#import-order-template

It's certainly supposed to enforce it[1] is a trivial example.

I had a quick play in cinder/quota.py and things are not doing what I'd expect.

Digging a little deeper it seems that hacking doesn't implement the template order [2] it only really enforces alphabetical.

So I guess it's okay to take this.

[1] http://paste.openstack.org/show/598740/
[2] http://git.openstack.org/cgit/openstack-dev/hacking/tree/hacking/checks/imports.py#n80

Patch-set: 2
Label: Code-Review=+2