Update patch set 2
Patch Set 2:

Btw, there is a proposed fix for 'pip' instead in https://github.com/pypa/pip/pull/4294 but until this is merged and released as a proper pip update I believe that it still makes sense to restrict the problematic setuptools versions for now. Not everybody will necessary get the latest pip version when installing any of the openstack components.

Patch-set: 2