Bump stevedore to 1.20.0
Stevedore 1.20.0 was recently released [1] and contains
a change we are waiting for neutron [2].

While the upper constraints have been bumped to 1.20.0
for stevedore, version 1.20.0 cannot be used by projects
and will fail [3] due to incompatible version.

This patch bumps global requirements for stevedore 1.20.0
so we can consume it in other project (ex: [2]).

[1] https://pypi.python.org/pypi/stevedore/1.20.0
[2] https://review.openstack.org/#/c/429915/
[3] http://logs.openstack.org/15/429915/3/check/gate-neutron-requirements/8f9f5ee/console.html

Change-Id: I5babd3025c63e5d3a9fac26a169278d849158bf9