Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

> Doug thanks.
 > I'm simply trying to use stevedore 1.20.0 in neutron [1] as
 > indicated in the commit message. As shown in [1], the
 > gate-neutron-requirements is failing (as noted in commit message).
 > 
 > 
 > I'm likely missing something very obvious; what do I need to wait
 > for before I can use stevedore 1.20.0?
 > 
 > 
 > [1] https://review.openstack.org/#/c/429915/

OK, you're close but not quite right on the process.

The first step is to update the minimum version here in this repo to indicate that we require at least 1.20.0 (see inline comment).

Then you can update neutron, as you have in the other patch. You can include the minimum version there or, as I describe in the comment on that patch, you can wait for the bot to propose an update.

I'll change my -1 to a -2 so that when you update this patch it can merge if another reviewer catches the update before I get back to it.

Patch-set: 1
Label: Code-Review=-1