Add virtualbmc to requirements and constraints
virtualbmc is a library/tool that is used throughout projects under
OpenStack Baremetal program for testing on virtualized hardware.

It depends on pyghmi library (pure Python IPMI protocol implementation)
which is already in g-r and u-c.

In order to avoid incompatible versions of virtualbmc and pyghmi
installed or required, virtualbmc should also be present in g-r and u-c.

Versions specified in this patch are latest currently available and are
known to work on CI already.

Change-Id: I72a170a0ab44673dd3041fe102d3b910e7db203f