Add virtualbmc to requirements and constraints
virtualbmc is a library/tool that is used throughout projects under
OpenStack Baremetal program for testing on virtualized hardware.

It depends on pyghmi library (pure Python IPMI protocol implementation)
which is already in g-r and u-c.

In order to avoid incompatible versions of virtualbmc and pyghmi
installed or required, which was already seen (ML thread [0]),
virtualbmc should also be present in g-r and u-c.

Version specified in this patch is latest available at the time of
stable/mitaka branch creation, and is known to work with pyghmi version
present in g-r and u-c as of stable/mitaka.

[0] http://lists.openstack.org/pipermail/openstack-dev/2017-March/113152.html

Change-Id: I72a170a0ab44673dd3041fe102d3b910e7db203f