Update patch set 2
Patch Set 2:

Sean, I am not expecting it to be merged before the master does :)

Just need something to test against changes to ironic's devstack plugin that would actually enable using virtualbmc from g-r.

Also, as the versions in the patch are quite different in mitaka from those in the patch to master, should it actually be a cherry-pick?

Patch-set: 2