Update patch set 3
Patch Set 3:

> (Err... I don't know what exactly means 'managed' vs 'unmanaged'
 > means here :-/)

This context 'managed' means that networking-bgpvpn subscribes to updates from the requirements team and agrees to *NOT* add items to *requirements.txt that are not listed in global-requirements [1,2].
 
 > What I know:
 > - networking-bgpvpn has a dependency on networking-bagpipe
 > - networking-bgpvpn stable/newton depends on networking-bagpipe >=
 > 5.0.0
 > (5.0.0 is the release that networking-bagpipe did roughly in sync
 > with Newton)
 > - since the moment networking-bagpipe did a 6.0.0 (roughly in sync
 > with Ocata), the tests for networking-bgpvpn stable/newton are
 > getting this new version of the networking-bagpipe package
 > - but networking-bagpipe 6.0.0 does not work with neutron's Newton
 > - I tried to have networking-bgpvpn stable/newton depends on
 > "networking-bagpipe >= 5.0.0, <6", but this was rejected by the
 > requirements checking bot [1]

Okay this is the bottom line here.  You're unable to set this requirement because somehow you
have a requirement on an item that is *not* in global-requirements.txt.  This is pretty confusing.

Okay I understand now what's going on.  The newton branch wasn't managed by the requirements team and therefore it has a dependancy on networking-bagpipe which isn't reflected in the openstack/requirements repo (in either globla-requirements.txt or upper-constratints.txt)  Add to this the proposal-bot isn't aware that it should be generating updates for this repo means that networking-bgpvpn is out of sync.

So when you try to update the specification on networking-bagpipe you get a failure because that isn't allowed. 

The correct fix is 1) NOT do that checking on stable/mitaka or stable/newton.  I'll do that and post a link to the review here ; and then 2) Restore 439668 and merge once it passes the gate

NOTE it may need to be squashed with https://review.openstack.org/439726 I'm not sure.

 > This is was lead me to add this upper-constraint to
 > openstack/requirements, but I'm of course open to any other
 > solution.
 > 
 > [1] http://logs.openstack.org/68/439668/2/check/gate-networking-bgpvpn-requirements/89d17d3/console.html#_2017-03-01_17_21_02_404982



[1] http://git.openstack.org/cgit/openstack/requirements/tree/projects.txt#n86
[2] http://git.openstack.org/cgit/openstack-infra/project-config/tree/zuul/layout.yaml#n12149

Patch-set: 3