Update patch set 2
Patch Set 2:

@Dima I think we're at an impass.

It's totally fine for you to disagree with me and as I've already stated if you do the next step is for you to contact the legal team for a review of the license terms and that rethinkdb 2.3.0 is available under and determine if that's appropriate for use by OpenStack.

Feel free to CC me on that discussion.

Patch-set: 2