Update patch set 3
Patch Set 3: Code-Review-1

> this failure has nothing to do with anything:
 > 
 > File "/home/jenkins/workspace/gate-requirements-tox-py27-check-uc-ubuntu-xenial/.tox/py27-check-uc/local/lib/python2.7/site-packages/pkg_resources/__init__.py",
 > line 1577, in _get
 > with open(path, 'rb') as stream:
 > IOError: [Errno 2] No such file or directory: '/home/jenkins/workspace/gate-requirements-tox-py27-check-uc-ubuntu-xenial/.tox/py27-check-uc/lib/python2.7/site-packages/pyparsing-2.1.10.dist-info/METADATA'
 > 
 > 
 > trying again, seems like build environment needs to be cleared and
 > rebuilt

We'er tracking this and while we don't understand the root cause we're attempting to work around it in 441811  So until that merges rechecking will fail.

Thanks for looking at this Mike.

To get this in we'll need to remove the cap in global-requirements *and* at the same time add a heap of entries to upper-constratints-xfails.txt.  this will allow us to merge the cap removal and keep the constraint in the 1.0 family while we wait for the requirements sync to propagate out.

Then when we up the constraint we can remove the entries from upper-constraints-xfails.txt and the world shoudl be a happier place.

The big question is can we remove the cap altogether?  We can constraints now so no project will advance to 1.2.0 when it's available without co-ordination. and capping makes removing the cap somewhat awkward.

The -1 is because of propagating the cap concept.

Patch-set: 3
Label: Code-Review=-1