Update patch set 3
Patch Set 3:

> add a heap of entries to upper-constratints-xfails.txt.

can you be more specific what this "heap of entries" consists of?  Or is this something you're going to do?

I would suggest that bumping SQLAlchemy "major" versions is something that should happen at least annually and if I can up the velocity, every six months or so.   Plus I'd imagine there have to be other dependencies that get an occasional major-version-bump.   Is there a standard process for how this is supposed to be done?

Patch-set: 3