Update patch set 3
Patch Set 3:

> > add a heap of entries to upper-constratints-xfails.txt.
 > 
 > can you be more specific what this "heap of entries" consists of? 
 > Or is this something you're going to do?

It's something we can do if you're generally okay with the plan.  I don't know the exact number but it'll be approximately 1 per item in global-requirements.txt that is managed by the requirements team and also depends on SQLAlchemy.
 
 > I would suggest that bumping SQLAlchemy "major" versions is
 > something that should happen at least annually and if I can up the
 > velocity, every six months or so.   Plus I'd imagine there have to
 > be other dependencies that get an occasional major-version-bump.  
 > Is there a standard process for how this is supposed to be done?


Yup the std. process is to have a library>=$semver in global-requirements and rely on upper-constraints.txt  Then when the new $major comes out once we've seen that it can be used without breakage we just bump the entry in global-requirements.txt

capping on master just makes it much harder to move forward.

Patch-set: 3