raise the version of appdirs to match what virtualenv uses
We're trying to install incompatible versions of appdirs, and in the
process breaking when we try to toggle back and forth [1]. Raise the
version of appdirs used to match what virtualenv brings with it to start
so that we avoid the toggle.

[1] http://logs.openstack.org/86/445186/3/check/gate-tempest-dsvm-neutron-src-cliff-ubuntu-xenial/2f8c8c0/logs/devstacklog.txt.gz#_2017-03-16_17_55_48_873

See also Ia95bbb7e4dd3503bf87559a118c0b197151fde16 in stable/ocata.

Change-Id: If3ad26b5450728c55a81e6e8d96a12bf0511bdf8
Signed-off-by: Doug Hellmann <doug@doughellmann.com>