Update patch set 2
Patch Set 2:

looks like this brings in grpcio and locks it to a specific version.  It also looks like grpcio is c based, which we generally prefer it not be I think.  It doesn't look like grpcio has additional deps, so it's just grpcio itself that needs to be added to gr and uc here.  But we (requirements team) should discuss if we want this in.  My two major gripes are as follows.

I don't really see testing for grpcio (could be missing it).

grpcio is C based (not the biggest concern to me, but annoying for others possibly)

grpcio is depended on with a specific version (=== in reqs.txt for etcd3)

Patch-set: 2
Reviewer: Gerrit User 14288 <14288@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0