Update patch set 2
Patch Set 2: Code-Review-1

Unfortunately, etcd3/grpc is not compatible with any async eventloop-based library (eventlet, gevent, or asyncio). I tested locally in os-lively using both gevent and eventlet, and found the following to be true:

1) You can use monkey_patch(all=True, thread=False) before importing your grpc code (i.e. import etcd3) and you stick to using threading.Thread().start(), reads/writes will work OK.

2) If you attempt to interact with etcd3 within an eventlet.spawn() or gevent.spawn().start(), you will hang.

Relevant issues in grpc and gevent projects in upstream:

https://github.com/grpc/grpc/issues/6046
https://github.com/grpc/grpc/issues/4629
https://github.com/gevent/gevent/issues/938
https://github.com/gevent/gevent/issues/786

So, we need to keep a watch (pun intended) on the above issues and see if there's any movement on them before making etcd3 a requirement for any project that uses eventlet/gevent/asyncio.

Best,

-jay

Patch-set: 2