raise the version of pyparsing to match what virtualenv uses
We're trying to install incompatible versions of pyparsing, and in the
process breaking when we try to toggle back and forth [1]. Raise the
version of pyparsing used to match what virtualenv brings with it to
start so that we avoid the toggle.

[1]http://logs.openstack.org/99/446899/1/check/gate-tempest-dsvm-neutron-src-oslo.config-ubuntu-xenial-ocata/b9b5c0c/logs/devstacklog.txt.gz#_2017-03-20_18_01_57_381

See also I2edbaa4cf8d7b4570473f7bce1f30fac503192a7 in stable/ocata.

Change-Id: If81430e4ee417271f365a6b7fd30e82ac10dd850