Revert "Update eventlet uc to 0.20.1"
This reverts commit a70d2236f0ea02890f47d1f9dd3b99e88bd486a8.

The Python 2.7 subprocess code [0] has a check for the presense
of select.poll() on import, but line 1447 does not use this
flag and attempts to call select.poll() anyway.

This is causing job failures, at least all Cinder docs jobs
are now failing because of this.

[0] https://svn.python.org/projects/python/trunk/Lib/subprocess.py

Additionally, this breaks designate non-docs jobs due to bundled
dnspython and interactions that causes. There is a patch landed upstream
to eventlet that will be in 0.21 that fixes the designate breaks. Let's
wait til that is released.

Change-Id: Icef3a281a065e34183be42d89dedb41d4be85aef
Closes-bug: #1674687