Add munch and jmsepath
These two libraries are used by shade, so in order for shade to sync
with global-requirements, we need to add them.

jmsepath is also a dependency of ansible, so anyone using ansible for
things already should already have it installed.

munch is a very lightweight library and is what shade returns from its
API instead of dicts of custom objects.

Change-Id: I4dd6be7a31d16a430c48ed4738fac4769c65975a