Update patch set 1
Patch Set 1:

the fix (in the code that fixings the '__version__ attribute error') that works in 0.9.5 should be backported to the openstacksdk stable/mitaka branch.  Once that code is backported a release of openstacksdk is made from the stable/mitaka branch (probably 0.8.2).  Once that is done we can update JUST the upper-constraint to 0.8.2.  If you also wish to mask known bad versions you can do so in global-requirements, but not by raising the minimum (we don't do that on stable branches).  >=0.8.0,!=0.8.1 or something.

Patch-set: 1