Update patch set 2
Patch Set 2:

Okay, there is a sync job but it isn't working because we had a non-g-r requirement in instack-undercloud, left over from before it was upstream.  I've proposed https://review.openstack.org/450902 which should get us back in sync and allow the bot to propose again.  I think that will get this patch working as well.

Patch-set: 2