Update patch set 1
Patch Set 1:

Tony, sorry I missed this as I was quite busy the last few days. Basically this alters the environment that is being installed for check-uc to install the upper-constraints version of packages (and hence it is checking those sets of packages) rather than the latest releases. this was mostly an issue on stable/* branches where the latest pbr release (2.0.0) was conflicting with everything else on stable/* I am not sure it makes a huge difference for master as we're usually pretty close in following releases there but its imho consistent to propose it for all branches.

Patch-set: 1
Reviewer: Gerrit User 6593 <6593@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0