Add jmespath
jmsepath is a python library implementing JMESPath
(http://jmespath.org/) which is a query language for JSON. shade added
JMESPath filter support inspired by a similar filter in Ansible. Ansible
already has a dependency on jmespath, so it's not a particularly onerous
new requirement in the OpenStack ecosystem generally.

It's also with noting that Amazon supports JMESPath expressions in their
CLI tool and in several of their SDKs.

- Is the library actively maintained?

Yes. It's been around for quite a while now and made its most recent
release 18 days ago. It also seems to be maintained by Amazon.

- Is the library good code?

Yes. It has good test coverage and has CI, and it works quite well.

- Is the library python 3 compatible?

Yes.

- Is the library license compatible?

Yes, it is MIT licensed.

- Is the library already packaged in the distros we target (Ubuntu
  latest / Fedora latest)?

Yes, in ubuntu since Trusty, in Fedora since at least F23.

- Is the function of this library already covered by other libraries
  in ``global-requirements.txt``?

No, it's not.

- Is the library required for OpenStack project or related dev or
  infrastructure setup? (Answer to this should be Yes, of course)
  Which?

shade allows users to filter returned lists of objects via jmsepath
expressions.

Change-Id: I963a558f5e294521ad7b067ae95cc02d379cf73b