Update patch set 3
Patch Set 3: Code-Review-1

I'm negative on this for the following reasons

This is needed for nemesis, which is not in our projects.txt
  The nemesis readme is kinda just the default, docs link to nowhere.
  Code looks like a very basic skeleton
  Can't add the code for this project in it's current state.

As for the library itself...

The setup.py doesn't depend requirements.txt (not sure why it's there if it's not used).
Dev status (while not the best metric) still lists it as pre-alpha.
Stated support for python3.4 and 3.5 but not 2.7?
Codewise it looks like a very small wrapper around oslo.log.

My suggestion is to get this type of thing included as somthing within oslo.log itself or just 'vendor' it like the link you have to the nova requestlog.

Patch-set: 3
Label: Code-Review=-1