Update patch set 3
Abandoned

We will be marking Mitaka EOL soon, so we no longer need or want to update the dependencies. If you consider this change critical, please start a thread on the openstack-dev mailing list requesting that it be reopened and considered before we finish the EOL work.

Patch-set: 3
Status: abandoned