Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Please answer the questions from https://github.com/openstack/requirements/#for-new-requirements in the commit message.

Further, the setup.py doesn't say it supports 3.5 (though it likely does).  It does not have testing, though is not the biggest library.  It also has not been updated in a while, though it could just be 'complete'.

Patch-set: 1
Label: Code-Review=-1