Update patch set 6
Patch Set 6: Code-Review+2

This looks like it's updating the client to the latest version on the branch, which seems OK. I'll leave the final approval for the stable team.

It's not clear where the API break is. It seems like that should have been reverted, though, rather than forcing users to update their client library.

Patch-set: 6
Label: Code-Review=+2