add ExaBGP to global-requirements.txt
This change adds ExaBGP [0] to global-requirements.txt.

Here are answers to the acceptability questions listed in [1]:

* Is the library actively maintained?

Yes, see for instance [1] and [2].

* Is the library good code?

In terms of codestyle its different than openstack (not pep8), but the code
is fine. In particular ExaBGP has reasonable testing baked in (unit tests and
functional BGP interop tests).

* Is the library python 3 compatible?

ExaBGP 4.0.0 is python2/python3 compatible.

* Is the library license compatible?

Yes (BSD).

* Is the library already packaged in the distros we target
(Ubuntu latest / Fedora latest)?

ExaBGP is packaged in Debian, Ubuntu. Fedora packaging is in progress [4].

* Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?

The consuming project is networking-bagpipe, which uses
ExaBGP as a library for BGP message encoding/decoding
(which is a subset of ExaBGP functionality).

Three years ago, bagpipe-bgp, one of the components of networking-bagpipe
was initially opensource outside the openstack community, and was carrying
an embedded fork of ExaBGP. To benefit from a uniformed contribution
workflow and solid/consistent CI, this code was moved in networking-bagpipe
two months ago, and the embedded ExaBGP fork removed in favor of a clean
dependency on upstream code, leading to this request for adopting ExaBGP
as an openstack dependency.

* Is the function of this library already covered by other libraries in
global-requirements.txt?

When the consuming project (networking-bagpipe/bagpipe-bgp) was started,
there was no equivalent to ExaBGP in Openstack requirements. In parallel
to the process of moving code from github to OpenStack (see above), Ryu [5]
was added to Openstack global-requirements.txt, but was not initially
supporting a range of BGP messages comparable to what ExaBGP was supporting
and not sufficient for networking-bagpipe. Today, Ryu seems to have
progressed, but the interfaces are different and we can't easily
do a forklift change to use Ryu in networking-bagpipe.

So, although the answer to the above question is 'yes', we hope that
the story leading to networking-bagpipe's dependency on ExaBGP can be
understood and should not prevent the project from doing the Right Thing
(moving code to the Openstack umbrella for a consistent CI and
contributor workflow, and avoiding "vendoring" a fork of ExaBGP).

[0] https://github.com/Exa-Networks/exabgp/
[1] https://github.com/openstack/requirements#for-new-requirements
[2] https://github.com/Exa-Networks/exabgp/commits/master
[3] https://github.com/Exa-Networks/exabgp/issues
[4] https://bugzilla.redhat.com/show_bug.cgi?id=1452649
[5] https://osrg.github.io/ryu/

Change-Id: I7d32246bdc0db40efa0c4b9b426c77f583063d86