Update patch set 1
Patch Set 1: Code-Review-1

Did you run into a build failure, or are you doing this for completeness?

We only want to maintain a list of packages that need to be co-installable here. The specs repos shouldn't be syncing requirements, limiting themselves to things in this list, or being co-installed with any of the application software.

Patch-set: 1
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1