Update patch set 1
Patch Set 1:

> Did you run into a build failure, or are you doing this for
 > completeness?
 > 
 > We only want to maintain a list of packages that need to be
 > co-installable here. The specs repos shouldn't be syncing
 > requirements, limiting themselves to things in this list, or being
 > co-installed with any of the application software.

This (and related patches) actually stems from a build failure in the specs repo caused by a lack of requirement sync, which can obviously be annoying from time to time to track down. Is it prescribed that requirements for specs repo be maintained manually? I looked at the projects.txt file to see there was another specs project in the list and thought it would be a nice thing to do to make sure requirements be kept in sync automatically. I believe it should be possible to have extra requirements in repos and have the bot ignore those, but while I was at it, I imaged it would be good to potentially address another chance of failure.

Patch-set: 1
Reviewer: Gerrit User 748 <748@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0