Update patch set 2
Patch Set 2: Code-Review-1

So, to clarify things, the issues.

1. The licence in the files are incompatable.
   https://github.com/rethinkdb/rethinkdb/blob/v2.3.0/drivers/python/rethinkdb/query.py#L1
   all rights reserved (in a few places)

This is on top of https://github.com/rethinkdb/rethinkdb/blob/v2.3.0/COPYRIGHT  which seems wrong

This software would need a release addressing these licence/copyright issues.

If we don't respond have a response in a week I think this is fine to abandon.  If in the future, past that point, the licence/copyright gets corrected then we can revisit this.

Patch-set: 2
Label: Code-Review=-1