Update patch set 1
Patch Set 1:

Doug,

no gates are broken exactly now, but these backports are to guard against possible failures in the future.

As per commit message, we already had troubles in stable/mitaka branch due to newer-than-needed virtualbmc installed. However, in Mitaka only a small number of ironic and its subrojects gate jobs were using it. Since Newton almost all Ironic etc gates involving actual node provisioning and DevStack do depend on virtualmbc, thus if similar breakage occurs in the future it will be much more severe (and this branches are to stay for another 1+ year).

Forgetting to add virtualbmc to g-r and u-c right from the start (or at least when it became crucial for ironic testing) was a mistake of ironic community and we sincerely apologize for that. We do need these backports to safely continue developing virtualbmc and making new releases of it when needed.

Patch-set: 1