Update patch set 2
Patch Set 2: Code-Review-1

Checking the deps of this.  It doesn't look like it depends on anything other than setuptools, however, they do have a requirements.txt that has ipaddr, which is not tracked by us yet.

They python classifiers only defines python/python2 which is a concern (I know you mentioned it worked with python3 though).

It's vendoring a bunch of python3 stuff?  This has me the most concerned. and the reason for the -1.

Beyond that I can't find the tests, unless it's the stuff under the qa directory.

Patch-set: 2
Label: Code-Review=-1