Update patch set 3
Patch Set 3:

we would have to fix docs and specs in several repos and in the case of the intel nfv ci also fix docs in openvswitch. i am not sure how much work this would be but it should be done even if this is merged but i will not have time to work on that.

i can fix the intel nfv ci by downgrading sphinx in our ci or cangaing how we compile ovs to not build docs but for the neutron gate someone would have to submit a patch to fix all broken docs.
similarly for the *-specs repos.

upstream ovs would also need to be eventually upudated but that is out of scope of this discussion.

Patch-set: 3
Reviewer: Gerrit User 11604 <11604@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0