Update patch set 3
Patch Set 3:

The ovs fix was already merged and downstreamed in gentoo at least.

https://github.com/gentoo/gentoo/commit/3ebea21b30582a7b11f1e17717b106ea0c660ec6

Just waiting on a release for them, ovs had a cve become public, so that should hopefully be soon.

If we can fix it in other places then the ovs fix should be the remaining piece right?

Patch-set: 3