Require oslo.messaging>=5.25.0
Change Ic6af07b98ff43806c2af38a3ba129991f1e0ec86 deprecated
the get_transport() method in favor of using get_rpc_transport()
but to use get_rpc_transport() we need to require a minimum
version of oslo.messaging that provides it. It was technically
available in 5.24.2 but since we already gate on 5.25.0 this just
uses that as the new minimum required version.

Change-Id: If2a790e1ee14c508ece697ba55c54d09f906cb3b