Update patch set 3
Patch Set 3: Code-Review-1

Just a question, have you reached out to the ryu contributors to see whether and what can be done to make a potential migration to ryu easier for both sides? 

For now I agree with Matthew's -1 that we should evaluate the options of avoiding a double dependency when there isn't a clear commitment to avoid the duplication going forward. Don't consider that a strong vote at this point in time, I'm missing most of the technical background of this. I'd be fine for a transitional amount of time between openstack releases to adopt the duplicate dependency if there is a clear path forward cleaning the duplication up.

Patch-set: 3