Pin six to 1.9.0
Enable various teams to change imports.

The siz library had the following changes:

The urllib, urllib2, and urlparse modules have been combined in the
urllib package in Python 3. The six.moves.urllib package is a
version-independent location for this functionality; its structure
mimics the structure of the Python 3 urllib package.

Resolves bug: #1696403

Change-Id: I12aa4c11e8ca0b4a8cb51f19d19aba4346599597