Update patch set 3
Patch Set 3: Code-Review+2

IMO this is tough call.  We have a library that seems adequate but includes vendored libraries which isn't great but it also isn't a violation of anything we've stated explicitly.

We also have a major but not complete overlap with another library (ryu) which *is* in global-requirements.

We also have networking-bagpipe working around ExaBGP being missing from global-requirements with [1].

Ideally I'd like to see the missing features added to ryu and networking-babgpipe utilizing  that library, both of which are massive hunks of work in the mean time we're likely to release networking-bagpipe with hidden dependencies which, I think, will make packaging networking-bagpipe hard er than it needs to me.

So I'm erring on the side of accepting this, even though it isn't obviously correct.

[1] http://git.openstack.org/cgit/openstack/networking-bagpipe/tree/tools/tox_install.sh#n65

Patch-set: 3
Label: Code-Review=+2