Update patch set 1
Patch Set 1: Code-Review-1

This is  reasonable as it's clearly a regression caused by [1].

When updating to also exclude pysnmp please also summarise the issue in the gate and link to the github issue.  If you don't report it upstream then we'll be effectively capping pyasn1 which isn't allowed.


[1] https://github.com/etingof/pyasn1/commit/679408903afee6d76b682376aedf0eeee496b627

Patch-set: 1
Label: Code-Review=-1