Update patch set 3
Patch Set 3: -Code-Review

The issue I see here is that while you can locally workaround the g-r check for one particular dependency, adding it to global-requirements is a wildcard approval for all other projects to be able to adopt it as well. 

If there isn't an integration issue with other projects then I'd still be tempted to leave it out of g-r for now.

Patch-set: 3
Label: Code-Review=0