Add tripleoclient to global requirements
Is the library actively maintained?
Yes.

Is the library good code?
Yes.

Is the library python 3 compatible?
Yes.

Is the library license compatible?
Yes. Apache 2.0

Is the library already packaged in the distros we target?
Fedora/CentOS/RHEL currently. Not currently available on Ubuntu.

Is the function of this library already covered by other libraries
in global-requirements.txt?
No.

Is the library required for OpenStack project or related dev or
infrastructure setup?
TripleO. tripleoclient is an openstack client plugin to control a
TripleO deployment. As it provides some functionality related to the
deployment of TripleO, we need to have this added so we can reference
it in testing the tripleo-heat-templates.
https://review.openstack.org/#/c/472731/

If the library release is managed by the Openstack release process does
it use the cycle-with-intermediary release type?
No it's currently cycle-trailing but we are not using pre-release
versions.

Change-Id: I895ffb811003859207ef1c8fdf44f327125b8bdf